+++ bin/build.sh
@@ -15,6 +15,10 @@ else
   echo "** building parsley.extend.min.js version " $1
   ruby ./bin/minify parsley.extend.js dist/parsley.extend.min.js $1 --force
   echo "  done!"
+
+  echo "** building parsley.es.min.js version " $1
+  ruby ./bin/minify l10n/parsley.es.js dist/parsley.es.min.js $1 --force
+  echo "  done!"
 fi
 
 if [[ "$2" == "doc" ]]